refactor: splitting ode_solver script
I had enough to scroll down to the bottom of the script.
