+++ bioptim/__init__.py
@@ -178,8 +178,8 @@ from .dynamics.fatigue.michaud_fatigue import MichaudFatigue, MichaudTauFatigue
 from .dynamics.fatigue.michaud_fatigue import MichaudFatigue, MichaudTauFatigue
 from .dynamics.fatigue.xia_fatigue import XiaFatigue, XiaTauFatigue, XiaFatigueStabilized
 from .dynamics.fatigue.xia_fatigue import XiaFatigue, XiaTauFatigue, XiaFatigueStabilized
-from .dynamics.ode_solver import OdeSolver, OdeSolverBase
-from .dynamics.ode_solver import OdeSolver, OdeSolverBase
+from .dynamics.ode_solvers import OdeSolver, OdeSolverBase
+from .dynamics.ode_solvers import OdeSolver, OdeSolverBase
 from .gui.online_callback_server import PlottingServer
 from .gui.online_callback_server import PlottingServer
 from .gui.plot import CustomPlot
